Refactored the file names and interfaces
